Use google places instead of geocode
Modified code to use google.maps.places.AutocompleteService and then
look up the place info (name and address) using getPlacePredictions.
This is a modified version of https://github.com/israelidanny/ion-google-place/pull/15/commits
which used an older version of this package. The places API search
functionality was added to this more up to date version.

If the there is a pre-existing location, this can be edited in the
popup. (The package didn't have this functionality built in!!!)

Also, the user can click the first item if the string entered does
not correspond to an existing place.